+++ app/services/stellar_base/account_subscriptions/get_operations.rb
@@ -17,7 +17,7 @@ module StellarBase
         c.operations = c.stellar_sdk_client
           .horizon
           .account(account_id: address)
-          .operations(order: "asc", limit: c.operation_limit)
+          .operations(order: "asc", cursor: c.cursor, limit: c.operation_limit)
           .records
           .map do |op|
             StellarOperation.new(op)