+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- AccountSubscriptions properly use cursor
+
 ## [1.1.0] - 2018-12-14
 ### Added
 - Added `StellarOperation` and `StellarTransaction` models